refactor: extract auto-stake-it flows
